feat: adding better tooltip to elements

+++ src/app/components/custom/table/table.component.ts
@@ -87,19 +87,59 @@ export class TableComponent<T extends TableItem> implements OnInit, OnChanges {
 
   initIcons(): void {
     const { hasCopy, hasDelete, hasEdit, hasView } = this.config;
-    const icons = [
-      !!hasCopy &&
-        IconConfig.withUrlType('DuplicateIcon', Icons.ContentCopy, DetailsTypes.Duplicate),
-      !!hasEdit && IconConfig.withUrlType('EditIcon', Icons.Edit, DetailsTypes.Edit),
-      !!hasView && IconConfig.withUrlType('ViewIcon', Icons.PageView, DetailsTypes.View),
-      !!hasDelete &&
-        IconConfig.withClick(
+    // const icons = [
+    //   !!hasCopy && {
+    //     ...IconConfig.withUrlType('DuplicateIcon', Icons.ContentCopy, DetailsTypes.Duplicate),
+    //     title: 'Duplicate',
+    //   },
+    //   !!hasEdit && {
+    //     ...IconConfig.withUrlType('EditIcon', Icons.Edit, DetailsTypes.Edit),
+    //     title: 'Edit',
+    //   },
+    //   !!hasView && {
+    //     ...IconConfig.withUrlType('ViewIcon', Icons.PageView, DetailsTypes.View),
+    //     title: 'View',
+    //   },
+    //   !!hasDelete && {
+    //     ...IconConfig.withClick(
+    //       'DeleteIcon',
+    //       Icons.Delete,
+    //       (id: number) => this.deleteItem.emit(id),
+    //       'warn'
+    //     ),
+    //     title: 'Delete',
+    //   },
+    // ].filter((i): i is IconConfig<number> => !!i);
+    const icons = [];
+    if (hasCopy) {
+      icons.push({
+        ...IconConfig.withUrlType('DuplicateIcon', Icons.ContentCopy, DetailsTypes.Duplicate),
+        title: 'Duplicate',
+      });
+    }
+    if (hasEdit) {
+      icons.push({
+        ...IconConfig.withUrlType('EditIcon', Icons.Edit, DetailsTypes.Edit),
+        title: 'Edit',
+      });
+    }
+    if (hasView) {
+      icons.push({
+        ...IconConfig.withUrlType('ViewIcon', Icons.PageView, DetailsTypes.View),
+        title: 'View',
+      });
+    }
+    if (hasDelete) {
+      icons.push({
+        ...IconConfig.withClick(
           'DeleteIcon',
           Icons.Delete,
           (id: number) => this.deleteItem.emit(id),
           'warn'
         ),
-    ].filter((i): i is IconConfig<number> => !!i);
+        title: 'Delete',
+      });
+    }
 
     this.icons = icons;
   }
